Move tests out of source files.
